move `FeatureVersion` and `FeatureVersions` into dedicated file; add test
